skidka_bot/requirements.txt
aioschedule==0.5.2
requests==2.28.1
pandas==1.5.1
aiogram==2.22.2
ydb==2.10.0
yandexcloud==0.190.0
protobuf==3.20.3
python-dotenv==0.21.0